requirements.txt
numpy
opencv-python
cryptography
pillow
tk
